implement GalleryPill with instrumentation
